Create Simple Code to automate repetitive task at work.
I work part-time at UPS as an Operations Manager while in university. Each night before leaving, I have to print out a selection of 'Call Tags' for the people in the morning to give to the drivers. It is not the most ideal way to code this process (uses mouse cords and other things that would not be universal.)  It does the trick for me though, would love to find ways to make this program more official and more universally sufficient.