Create mouse coordinates
Simplest code ever. Once running you have a second or two to move your mouse to the location you want to know the coordinates for, then it prints out the coordinates for you. I used this when automating my other program, soon I hope to learn a way to avoid having to use this technique, however.